fix: address DynamoDB review feedback

Copilot-Session: 30769b52-5dfb-4bc6-9cb9-6bf16f79c3f4

+++ src/AWS/Orleans.Transactions.DynamoDB/TransactionalState/StateEntity.cs
@@ -44,7 +44,7 @@ internal class StateEntity
             this.State = state.B.ToArray();
 
         if (fields.TryGetValue(DynamoDBTransactionalStateConstants.ETAG_PROPERTY_NAME, out var etag))
-            this.ETag = int.Parse(etag.N);
+            this.ETag = long.Parse(etag.N);
     }
 
     public static string MakeRowKey(long sequenceId)
